Move profile parsing to a trait
